Shift to python-gpg library
The python-gnupg library is a wrapper over gpg command, it is nice and
and simple to use but it is rather limited (ex: no subkey management, no
access to keygrip, etc). The code has been migrated to python-gpg, a
binding over GnuPG library, a bit more complicated but full of features.