Dbus activation
There is a new timer thread that tracks timestamp of the last activity.
When this timer is over (after 30 seconds of inactivity), the service
is automatically stopped.

The service manager is both notified for watchdog and when the service
stops, so it handles the new requests on DBus before the service is
restarted.