Support server side pre-modified tarballs
If a script pre.sh is present in the artefact definition, extract the
upstream tarball and execute this script (using bash shell) in the root
of the tarball. The purpose of this script is to modify the upstream
tarball (eg. add vendored libraries not available in the distribution)
to make integration easier. Then, a new tarball is generated for the
package build.

A new ArtefactBuild attribute tarball is introduced, it is the final
pathlib Path to the upstream tarball, modified by pre-script or not.

A new source variable is introduced in the RPM spec file template with
the name of the final archive used for the build, modified or not.