Client uses generic server URI
Instead of a host/instance pair, the clients (HTTP and Dbus) now takes
a URI to define server endpoint. Fatbuildrctl does not have the notion
of instance anymore, it just connects to a server URI which can contain
the instance name, or not.

This notably makes fatbuildrctl able to connect to fatbuildrweb
launched on a specific instance, without the instance names in the API
URLs.