Make task submission code generic
Tasks names are used at Dbus FatbuildrInterface level to submit tasks.
This makes possible to remove a lot of duplicated code for all type of
tasks at Dbus FatbuildrMultiplexer and ServerTasksManager levels.

The ProtocolRegistry now saves the caller to initialize the tasks,
defaulting to the type of the RunningTask by default. This permits to
use the factory for ArtefactBuild.

The Submit() Dbus method has been renamed Build() along with the
wrappers on client side, to avoid confusion between build submission and
generic tasks submissions.