Support pre-built source tarball
The fatbuildrctl build command is now able to generate a source tarball,
using --source-dir and --source-version options. The generated source
tarball is sent within the build request and used as the upstream
original tarball for this build.

This feature can be useful to build packages of unreleased versions of
software.