Support multi-arch
Supported architectures can be declared as a list in instances
pipelines configuration file. By default, the host architecture is
used. The host architecture is automatically added to the list of
supported architecture, even if it is not present in pipelines
configuration file.

One build environment per supported architecture is maintained in
images.

When user submit artefact build, fatbuildrctl checks if the artefact is
architecture dependent, checking respective parameters in deb/control or
in RPM spec file. If not, the build is submitted to the first found
architecture. Otherwise, the build is submitted for all supported
architectures in the pipelines.

New specifics module is introduced to abstract architectures names
differences between distributions/formats, so users only manipulate
normalized architectures names.

Builds environments and artefact builds with architectures not natively
supported by the host are run with qemu-user-static, thanks to binfmt
registrations. This requires qemu-user-static to be installed both in
the images, thus modifications in *.mkosi files, and on the host for
registrations at kernel level.