Fix multi-arch deb registry requests
Reprepro has a distinction between repository architecture (ie. the
location in the repository) and package architecture. The source
packages must be checked against the repository architecture while
binary packages must be checked against the package architecture.

We also ensure noarch packages are not duplicated in resulting list
as reprepro find them for every supported architectures in the
repository.