Update deb packaging for fatbuildr system user
An additional architecture independent package fatbuildr-common is
introduced to create the fatbuildr system user, using the sysusers.d
configuration file and calling systemd-sysusers in postinst.

Another package fatbuildr-wrappers is introducced, with architecture
any. It contains wrappers for systemd-nspawn and mkosi with execution
restricted to fatbuildr system user. The permissions on the wrappers are
setup in postinst script.

The main fatbuildr package now depends on new -common and -wrappers
packages.

The debian/dirs is largely cleared as fatbuildrd state/cache
directories, including their subdirectories, are now created by systemd
when the service is started.

The debian/rules is modified to remove the dh_systemdinstall override
which became useless. Build and install rules are overriden to execute
the wrappers Makefile, additionally to Python setup.py.