Support dist release tag in InstancePipelines
The format of distributions declaration in pipelines definitions
evolves, it becomes a list with 3 keys: name, env and tag.

For example:

  formats:
    deb:
      bullseye: bullseye
      bookworm: bookworm
    rpm:
      el8: rocky-8

Becomes:

  formats:
    deb:
      - name: bullseye
        env: bullseye
        tag: deb11
      - name: bookworm:
        env: bookworm
        tag: deb12
    rpm:
      - name: el8
        env: rocky-8
        tag: el8

This way, it becomes possible to define distribution release tags to
append to artifacts release numbers for all distributions. It also makes
the purpose of the fields clearer.