Add prescript feature for supplementary tarballs
It is now possible to declare #PRESCRIPT_TARBALLS token in prescript
followed by a space separated list of subdirectories. For each of these
subdirectorie, fatbuildrd generates a supplementary tarball that will
be used to build the artifact source tree along the upstream main
tarball.

This is an alternative to the patch generation, which scales better for
large files and support binary blobs.

In RPM spec file, the purpose of source token in templates has changed,
it now holds the declaration of all artifact source archives, including
potential supplementary tarball generated by the prescript. New
prep_sources token is also introduced with appropriate %setup command
in %prep section.