Give supplementary tarball unique names
Supplementary tarballs generated with prescript are now renamed with
timestamp and part of task id to avoid conflict in package repository
with potentially existing supplementary tarballs with the same name
but different content generated by previous build of the same artifact.

Symlink are created with the generic name targetting the uniquely named
subdirectory so the packaging code can use the generic name and does
not have to cope with these unique names. The symlinks are installed in
source package with a generated patch appended in artifact patch queue.