Publish one DBus object per instance
Instance of one DBus object and its interface for all instances,
fatbuildrd now publishes now DBus object per running instance. Each
instance has its name on the bus, and methods signature in the API is
simplified as it does not have need the instance argument for every
basically every calls like before.

Note that dasbus DBusContainer are not used to publish objects as it
does not let control over the object names on the bus, they are
dynamically generated using continuously incrementing integer counter,
which is not ideal when introspecting the bus. With this commit,
instances objects are published with their name (ex: default) in the
dedicated Instances namespace.