Add conf param to control seccomp on containers
This commit introduces the seccomp parameter in containers section of
system configuration file to control whether Linux seccomp is used to
filter the syscalls available to containers. Under the hood, it controls
whether the SYSTEMD_SECCOMP environment variable is defined for
systemd-nspawn.

The default value is yes (ie. seccomp is enabled). This should be the
desirable setting for most systems, unless seccomp does not work as
expected.