refactor: rename meta.yml to artifact.yml
The filename meta.yml is meaningless, it was not well chosen. The
purpose of this file is hopefully clearer after the rename to
artifact.yml. Note that artifact.yaml (with full yaml extension) is also
supported.

The previous filename meta.yml is still supported but the user is warned
with a deprecation notice.

The commit also updates the documentation and import-srcrpm utility
accordingly.

Fix #73