feat(cli): add images env-shell subcommand
This commit adds CLI fatbuildrctl images env-shell subcommand to open an
interactive shell in a build environment. The build environment must be
selected with --format, --distribution and --architecture option.

Configuration parameter shell_cmd is introduced in format:{deb,rpm}
sections to define the command to run in the format container to
instanciate this interactive shell in the build environment.

The documentation (manpage and configuration parameters) is updated
accordingly.