feat: add command to generate JWT token
This commit adds fatbuildrctl token command to generate a JWT token. The
token is generated by fatbuildrd daemon through DBus. It requires a new
manage-token Polkit permission which is attributed to all fatbuildr
system group members in default policy.

The daemon generates and saves an encryption key per instance at startup
time. The saved encryption key is reloaded on the following restarts.
This key is used for signing the generated JWT tokens.

The JWT token is generated in a new tokens modules that depends on PyJWT
external library.