feat(web): JWT token based authentication w/ RBAC
This commits adds to fatbuildrweb support of JWT token based
authentication. A RBAC policy can be defined to check for permissions
based on roles. Roles are defined with lists of individual users and
groups. A default policy is provided to restrict tasks submission to
members of fatbuildr group.

The documentation is heavily updated with configuration details.