feat(cli): support token based auth to REST API
This commit adds to fatbuildrctl command the support of JWT token based
authentication on Fatbuildr REST API. When the token is available at the
path defined in user preferences file or in default path
~/.local/share/fatbuildr/token, it is automatically send in HTTP
requests headers for authentication. The HTTP status codes sent by the
server are checked and permission errors are reported to users.