feat(cli): support multiple tokens
The fatbuildrctl client now supports loading multiple tokens in a
directory defined in user preferences file. The filename is the targeted
URI encoded in base64 with the token prefix. This way, users can have
multiple tokens to connect to multiple URIs, and only one token per
URI can be saved in directory.

The manpage is also updated to mention the new tokens parameter in the
user preferences file.