fix: avoid ~ removal in version through REST API
The tilde was removed from source tarball filenames by REST API server
for security reasons but the artifact main version number is extracted
from this filename, leading to version number with legitimate tilde
being modified. This is fixed by this commit. The security of this
filename is now ensured by a simple basename extraction.

Fix #81