refactor: common function to extract source tree
This commit introduces the new utility function
extract_artifact_sources_archives() to extract an artifact source tree
in a consistent way in all places where it is required.

In the same time, the archive module with classes to manipulate archives
and sources archives are introduced. They usefully replace the many
archives related functions in utils module.