fix: extract other sources in versioned dir
The supplementary sources are now extracted in the source tree in a
subdirectory named after the tarball filename instead of the generic
artifact source name. A symlink is added in the artifact packaging
sources (with a patch) from the generic artifact source name to the
versioned subdirectory to simplify package installation code.

This change notably allows naming debian source packages supplementary
sources with names changing with versions, thus avoiding conflict
between multiple archives that would have had the same name but
different checksums in the repository otherwise.