fix: extract supplementary source in sanitized dir
Unfortunately, Debian source packages only support supplementary source
tarballs with component name that matches [[:alnum:]-]+. Documentation
has not been found for this implementation detail but the regexp can be
found in dpkg source code:

  https://git.dpkg.org/cgit/dpkg/dpkg.git/tree/scripts/Dpkg/Source/Packa
  ge.pm#n406

To match this requirement, a sanitzed_stem property is added to
ArchiveFile class to removes all characters expect alphanumeric and
dash. It is then used to determine the name of the subdirectory where
supplementary source archives are extracted in artifact source tree.