fix: remove deprecated source RPM after build
After a successful build of RPM packages, the deprecated binary RPM
packages (ie. the previous version of the packages of the same artifact)
were removed from the repository but the source package was forgotten,
resulting in all versions of an artifact source RPM package living
permanently in registry. This is now fixed by this commit.

Fix #58