fix: systemd-nspawn exec error in OSI container
The mkosi utility executes systemd-nspawn container technology to
execute command in the built image, eg. to remove documentation. By
default, systemd-nspawn tries to connect to DBus system session to
register the container on systemd-machined and obtain a specific unit
dynamically. Unfortunately, this cannot work inside another container
where DBus system session is not available. This makes OSI builds fail
as soon as it has to launch systemd-nspawn.

The solution is to launch mkosi with --nspawn-keep-unit argument to make
it launch systemd-nspawn with option to avoid registering.

Fix #103