fix: unsupported GPG key for Debian OSI images
This commit fixes an error encountered when mkosi runs apt-get update
during the build of Debian images indicating that the keyring file
provided by debian-archive-keyring package has an unsupported filetype.
The error message is inaccurate as it was discovered with strace that
apt-get executes apt-key, which on its side tries to execute the cmp
command that was actually missing in OSI format container image. The cmp
command is installed with diffutils package (which is arguably an apt
package dependency).

Fix #104