fix: search mkosi output in distro~release subdir
Starting from version 14, mkosi systematically creates a subdirectory
named after the distribution and the release separated by tilde (~) in
the output directory.

For reference, see this commit:
https://github.com/systemd/mkosi/commit/b40f7d1

Fatbuildr now searches for this subdirectory using the glob pattern *~*
and expects to find in there the built images and SHA256SUMS file to
sign.

Fix #105