fix: permission error on OSI artifacts
To produce artifacts in OSI format, mkosi is run as root in container.
This makes the resulting files owned by root instead of the user running
fatbuildrd daemon. As a workaround to this, the SUDO_UID environment is
defined to fake sudo and make mkosi chown the content of the output
directory to this particular user eventually.

Fix #106