feat: give possibility to build OSI on host
A new containerized boolean parameter is introduced in system
configuration to control if the OSI build is run in container or
directly on host. This is useful for some OSI formats for which mkosi
requires access to loop devices that are generally unavailable in
containers because Linux kernel is not able to virtualize them, at least
for the moment.