feat: multiple OSI in the same checksum file
This commit adds support of multiple OSI images referenced in the
checksum file published in the same registry directory. The new OSI
files are added in the resulting checksum file, the checksums of the
updated OSI files are updated in the resulting checksum file and the
other entries in the previous checksum file are left intact.

The signature of the checksum file is now performed in RegistryOsi
instead of ArtifactBuildOsi, on the resulting merged checksum file
rather than on the checksum file generated by mkosi.

Fix #114