fix(dbus): retrieval of instances w/o default
Fatbuildrweb was retrieving the list of instances defined in fatbuildrd
by connecting of default instance through DBus. It could not work when
default instance is not defined on fatbuildrd. The solution is to
defined a new DBusServiceClient to connect without instance name to get
access to the instances list call.

The DBusClient has been renamed DBusInstanceClient to mark the tie of
this client with a specific instance.

Fix 127