fix(cli): HTTP streamed task output watching
When watching streamed tasks output on Fatbuildrweb HTTP REST API,
fatbuildrctl crashed due to unexpected stopped iterator. Actually, the
python requests Response.iter_content() was not properly used and with
recent change for better integration of urllib3, a regression occured.
The chunk size of the generator cannot be changed dynamically while
reading data. The data is now placed into a buffer, and the buffer is
filled with the next chunks when new data is required.

fix #138