fix(web): headers to avoid watch buffering
Reverse proxies usually perform some caching and buffering for
performance reasons between clients and upstream server. While this is
generally good in most cases, this is particularly annoying when wathing
tasks outputs as it introduces long latencies and lines of output in big
chunks. To avoid this behaviour, set X-Accel-Buffering HTTP response
headers with disabled value to instruct potential reverse proxy to avoid
buffering for this response.