feat: execute cmd with images env-shell
This commit adds the possibility to specify a command to execute in
build environment in arguments of fatbuildrctl images env-shell command.

New exec_cmd configuration parameter is introduced to specify the
command used to execute the user command in the build environment. Also,
the exec_tmpfile boolean parameter is added to specify if the user
command to execute must be written in a temporary script file for
execution by the exec_cmd command. This is notably required by
cowbuilder/pbuilder.

fix #98