fix(conf): support mkosi >= 15
The [images]>create_cmd is updated with arguments and values expected by
latest mkosi. Also, the images definitions are updated with some syntax
changes.

In latest versions of mkosi, postinstall script is executed outside the
chroot by default, as well as the finalize script. The postinstall
script are updated to reflect this change, by using mkosi-chroot
utility when required. The finalize script for RPM image is now merged
in the postinstall script, as it has no reason to be separated anymore.