fix(conf): use nspawn --resolv-conf=bind-stub
It has proven to be more reliable, less prone to errors with actual
image content, and this value is recommended in systemd-nspawn manpage
as soon as this file is not subject to change in the containers (which
is the case in Fatbuildr).

The apparmor profile is updated to allow this bind mount.